+++ tests/compare.spec.js
@@ -3,7 +3,7 @@ import { t_cmp } from './setup.js';
 import fc from 'fast-check';
 
 fc.configureGlobal({ numRuns: 300 });
-const N = 2 ** 12; // 2**31-1 max
+const N = 4096; // 2**31-1 max
 
 t.test('compareTo', async t => {
   fc.assert(