+++ tests/log.spec.js
@@ -1,28 +1,43 @@
 import t from 'tap';
-import { t_log2 } from './setup.js';
+import { t_log2, t_log10 } from './setup.js';
 import fc from 'fast-check';
 
 fc.configureGlobal({ numRuns: 300 });
 
-function log2(n) {
-  if (n < Number.MAX_SAFE_INTEGER) {
-    const x = Number(n);
-    return BigInt(Math.floor(Math.log2(x)));
-  }
-  return BigInt(n.toString(2).length - 1);
+const M = 2n ** 4096n;
+
+function log(k, n) {
+  return BigInt(n.toString(k).length - 1);
 }
 
 t.test('log2', t => {
   fc.assert(
-    fc.property(fc.bigInt({ min: 1n, max: 2n ** 4096n }), n => {
-      t.equal(t_log2(n), log2(n));
+    fc.property(fc.bigInt({ min: 1n, max: M }), n => {
+      t.equal(t_log2(n), log(2, n));
     })
   );
 
   // Identity
   fc.assert(
-    fc.property(fc.bigInt(1n, 2n ** 8n), n => {
+    fc.property(fc.bigInt(1n, 1000n), n => {
       t.equal(t_log2(2n ** n), n);
+    })
+  );
+
+  t.end();
+});
+
+t.test('log10', t => {
+  fc.assert(
+    fc.property(fc.bigInt({ min: 1n, max: M }), n => {
+      t.equal(t_log10(n), log(10, n));
+    })
+  );
+
+  // Identity
+  fc.assert(
+    fc.property(fc.bigInt(1n, 1000n), n => {
+      t.equal(t_log10(10n ** n), n);
     })
   );